+++ examples/resources/fmc_icmpv4_object/resource.tf
@@ -1,5 +1,5 @@
 resource "fmc_icmpv4_object" "example" {
-  icmp_type   = 3
+  icmp_type   = "3"
   code        = 0
   name        = "my_icmpv4_object"
   description = "ICMPv4 network unreachable response, type 3, code 0"