+++ examples/resources/fmc_icmpv4s/resource.tf
@@ -2,7 +2,7 @@ resource "fmc_icmpv4s" "example" {
   items = {
     my_icmpv4_objects = {
       description = "ICMPv4 network unreachable response, type 3, code 0"
-      icmp_type   = 3
+      icmp_type   = "3"
       code        = 0
     }
   }