+++ examples/resources/fmc_icmpv6_object/resource.tf
@@ -1,5 +1,5 @@
 resource "fmc_icmpv6_object" "example" {
-  icmp_type   = 1
+  icmp_type   = "1"
   code        = 3
   name        = "my_icmpv6_object"
   description = "ICMPv6 address unreachable response, type 1, code 3"