+++ internal/provider/data_source_fmc_icmpv4_object.go
@@ -72,12 +72,12 @@ func (d *ICMPv4ObjectDataSource) Schema(ctx context.Context, req datasource.Sche
 				MarkdownDescription: "Name of the FMC domain",
 				Optional:            true,
 			},
-			"icmp_type": schema.Int64Attribute{
-				MarkdownDescription: "ICMPv4 [type number](https://www.iana.org/assignments/icmp-parameters/icmp-parameters.xhtml).",
+			"icmp_type": schema.StringAttribute{
+				MarkdownDescription: "Type number, or `Any` for any type.",
 				Computed:            true,
 			},
 			"code": schema.Int64Attribute{
-				MarkdownDescription: "ICMPv4 [code number](https://www.iana.org/assignments/icmp-parameters/icmp-parameters.xhtml) subordinate to the given `icmp_type`.",
+				MarkdownDescription: "Code number subordinate to the given `icmp_type`.",
 				Computed:            true,
 			},
 			"name": schema.StringAttribute{