+++ internal/provider/data_source_fmc_icmpv6s_test.go
@@ -59,7 +59,7 @@ func testAccDataSourceFmcICMPv6sConfig() string {
 	config := `resource "fmc_icmpv6s" "test" {` + "\n"
 	config += `	items = { "my_icmpv6_objects" = {` + "\n"
 	config += `		description = "ICMPv6 address unreachable response, type 1, code 3"` + "\n"
-	config += `		icmp_type = 1` + "\n"
+	config += `		icmp_type = "1"` + "\n"
 	config += `		code = 3` + "\n"
 	config += `		overridable = true` + "\n"
 	config += `	}}` + "\n"