+++ internal/provider/resource_fmc_icmpv4_objects.go
@@ -104,15 +104,12 @@ func (r *ICMPv4ObjectsResource) Schema(ctx context.Context, req resource.SchemaR
 							MarkdownDescription: helpers.NewAttributeDescription("Indicates whether object values can be overridden.").String,
 							Optional:            true,
 						},
-						"icmp_type": schema.Int64Attribute{
-							MarkdownDescription: helpers.NewAttributeDescription("ICMPv4 [type number](https://www.iana.org/assignments/icmp-parameters/icmp-parameters.xhtml).").AddIntegerRangeDescription(0, 255).String,
+						"icmp_type": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Type number, or `Any` for any type.").String,
 							Optional:            true,
-							Validators: []validator.Int64{
-								int64validator.Between(0, 255),
-							},
 						},
 						"code": schema.Int64Attribute{
-							MarkdownDescription: helpers.NewAttributeDescription("ICMPv4 [code number](https://www.iana.org/assignments/icmp-parameters/icmp-parameters.xhtml) subordinate to the given `icmp_type`.").AddIntegerRangeDescription(0, 255).String,
+							MarkdownDescription: helpers.NewAttributeDescription("Code number subordinate to the given `icmp_type`.").AddIntegerRangeDescription(0, 255).String,
 							Optional:            true,
 							Validators: []validator.Int64{
 								int64validator.Between(0, 255),