+++ internal/provider/resource_fmc_icmpv4_test.go
@@ -82,7 +82,7 @@ func testAccFmcICMPv4Config_all() string {
 	config := `resource "fmc_icmpv4" "test" {` + "\n"
 	config += `	name = "my_icmpv4_object"` + "\n"
 	config += `	description = "ICMPv4 network unreachable response, type 3, code 0"` + "\n"
-	config += `	icmp_type = 3` + "\n"
+	config += `	icmp_type = "3"` + "\n"
 	config += `	code = 0` + "\n"
 	config += `	overridable = true` + "\n"
 	config += `}` + "\n"