+++ internal/provider/resource_fmc_icmpv6_test.go
@@ -82,7 +82,7 @@ func testAccFmcICMPv6Config_all() string {
 	config := `resource "fmc_icmpv6" "test" {` + "\n"
 	config += `	name = "my_icmpv6_object"` + "\n"
 	config += `	description = "ICMPv6 address unreachable response, type 1, code 3"` + "\n"
-	config += `	icmp_type = 1` + "\n"
+	config += `	icmp_type = "1"` + "\n"
 	config += `	code = 3` + "\n"
 	config += `	overridable = true` + "\n"
 	config += `}` + "\n"